Add character customization: shape sliders, schema-driven body, editor panel
A humanoid body whose shape is driven by component data, editable live in
the editor and bakeable at cook. Every characteristic is a morph target or
a per-joint proportion; nothing is bespoke per-slider vertex code.

- CharacterShape asset: named sliders resolve to morph targets by a
  `name` / `name+` / `name-` convention; per-joint proportions apply a
  uniform scale and a bone-length offset to the posed skeleton. A static
  morph base layer composes under animation morph tracks, and the
  character capsule follows the proportioned height.
- CharacterSchema + CharacterModel assets: the body contract (joints,
  keys, regions by dominant skin weight, proportion groups, panel
  sections, presets) is declarative data validated at cook, with a bundled
  humanoid schema. A model expands into a SkinnedMesh of the same name,
  so existing refs, animations, and the payload cache need no changes.
- Cook-time synthesized targets (girth, taper, bulge, mirror, blend_mask,
  surface_offset) generate 29 extra sliders from any conforming mesh.
  `bake: true` flattens a shape into a fixed mesh for NPCs. Skinned LODs
  come from the existing QEM index decimation.
- Editor CharacterShape panel with a drag-slider widget: rows derive from
  the schema, drags preview on the live pose without a rebuild and commit
  one undo step on release. Skinned meshes are now pick candidates and
  render shaped under the edit-mode overlay.
- Blender-authored base humanoid (18,970 verts, 25 bones, 21 shape keys,
  rotation-only idle) ships as the customize_character example.

Also: sprite outline strips for transparent fills, an autorelease pool
around parallel Metal pass encoding, and a SkinnedLods argument struct.

Store skinned-mesh morph targets sparse end to end

- Replace the dense MRPH payload block with MRPS: vertex-major CSR of
  u32 offsets + 28-byte MorphEntry (target, position, normal), deltas
  under 1e-6 dropped; new concinnity-cpu gfx::morph_targets module with
  dense<->sparse round trip, validation and shader layout test
- Cook sparsifies in compile_skinned_mesh_payload_with_lods; world JSON
  stays dense
- All three rt_skin kernels and Metal's legacy skinned VS read one packed
  offsets+entries buffer through the existing morph binding and walk only
  the entries touching their vertex
- base_humanoid blob 37.2 MB -> 11.8 MB; Metal and MoltenVK captures
  match the dense baseline; DirectX type-checked, Windows run owed

Add glTF export from the editor (character customization phase 6)

- New concinnity-editor gltf_export module: hand-rolled GLB writer
  (container, buffer packing, JSON emission), export-side CharacterShape
  bake, and a world-to-payload source that compiles the working entries
  in memory so exports carry the composed mesh with every morph target
- /export [name] [bake] console command plus an "Export .glb" row in the
  asset tree's row menu for skinned meshes; runs on a worker and writes
  <name>.glb beside the project with toast/log reporting
- Cook: PipelineResult::resource_payload(kind, name) accessor and an
  inline-form skinned glTF decode helper (import-side only)
- Tests: writer units, cook round-trip, real-body export, hook + panel
  coverage; verified by importing the export into Blender (all 70 shape
  keys by name, 25 bones)

+++ crates/concinnity-core/src/gfx/render_types.rs
@@ -1612,9 +1612,7 @@ pub struct SkinnedDrawObject {
     /// not consumed by the draw call itself, but the asset hot-reload's
     /// `rebuild_skinned_geometry` path needs to know each slot's vertex
     /// region to copy unchanged geometry across a size-changing rebuild.
-    /// Stored as u16 because the shared skinned index buffer is u16 and
-    /// every slot's `vertex_base` has to fit there.
-    pub vertex_base: u16,
+    pub vertex_base: u32,
     /// Number of vertices in this slot's region of the shared skinned
     /// vertex buffer.
     pub vertex_count: usize,
